Fixed missing data bug in ```RYcoder```
Fixed ```RYcoder``` bug where, if the input file (NEXUS or PHYLIP) contained missing bases coded as N's but PHYLIP output was desired (implicit, or called using -a flag), then the output PHYLIP would _also_ contain missing bases coded as N's. This would cause problems in downstream phylogenetic analyses, for example missing bases need to be coded as question marks in order for RAxML to read files correctly (otherwise throws error and quits). Fix was to reprocess the RY matrix to convert N's (if present, in either lowercase or uppercase) to question marks, '?'.