Updating FASTA2PHYLIP function
Updating function startup/organization steps to include conditional correction of FASTA filenames including "-" or " " (space) characters, which will cause the function to fail. These characters, if present in the input FASTA file names, are now converted to "_" (underscores). Expect the user to note this change in output, but consider adding code to print an INFO / WARNING message to screen.